zap clang warning
git-svn-id: https://svn.r-project.org/R/trunk@86155 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/eval.c
@@ -7181,7 +7181,7 @@ static SEXP bcEval(SEXP body, SEXP rho, Rboolean useCache)
       /* initialize binding cache on the stack */
       if (R_BCNodeStackTop + n + 1 > R_BCNodeStackEnd)
 	  nodeStackOverflow();
-      R_BCNodeStackTop->u.ival = n;
+      R_BCNodeStackTop->u.ival = (int)n;
       R_BCNodeStackTop->tag = CACHESZ_TAG;
       R_BCNodeStackTop++;
       vcache = R_BCNodeStackTop;
